getFriendInfoById方法。发送者？接收者？谁加密？谁解密？我懵啦，啦啦啦。

+++ app/src/main/java/com/example/endtoendencryptionsystem/entiy/dao/FriendsDao.kt
@@ -19,7 +19,7 @@ interface FriendsDao {
      * id:当前用户的id
      */
     @Query("select * from im_friend where friend_id=:id")
-    fun selectFriendsByFriendId(id: Int): Friend
+    fun selectFriendsByFriendId(id: Long): Friend
 
     /**
      * 添加好友
